requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4
pydantic==2.8.2
streamlit==1.37.1
beautifulsoup4
dateparser

